requirements.txt
fastapi
uvicorn[standard]
openai
aiosqlite
python-dotenv
jinja2
httpx
